+++ packages/presets/src/theme/tokens/colors.ts
@@ -3,16 +3,31 @@ import { defineTokens } from '@pandacss/dev'
 export const colors = defineTokens.colors({
   neutral: {
     25: { value: '#fdfdfd' },
-    50: { value: '#fafafa' },
-    100: { value: '#f5f5f5' },
-    200: { value: '#e5e5e5' },
-    300: { value: '#d4d4d4' },
-    400: { value: '#a3a3a3' },
-    500: { value: '#737373' },
-    600: { value: '#525252' },
-    700: { value: '#404040' },
-    800: { value: '#262626' },
-    900: { value: '#171717' },
-    950: { value: '#0a0a0a' },
+  },
+  stone: {
+    25: { value: '#fcfcfc' },
+  },
+  zinc: {
+    25: { value: '#fcfcfc' },
+  },
+  gray: {
+    25: { value: '#fcfcfd' },
+  },
+  slate: {
+    25: { value: '#fcfcfd' },
+  },
+  grayPalette: {
+    25: { value: '{colors.neutral.25}' },
+    50: { value: '{colors.neutral.50}' },
+    100: { value: '{colors.neutral.100}' },
+    200: { value: '{colors.neutral.200}' },
+    300: { value: '{colors.neutral.300}' },
+    400: { value: '{colors.neutral.400}' },
+    500: { value: '{colors.neutral.500}' },
+    600: { value: '{colors.neutral.600}' },
+    700: { value: '{colors.neutral.700}' },
+    800: { value: '{colors.neutral.800}' },
+    900: { value: '{colors.neutral.900}' },
+    950: { value: '{colors.neutral.950}' },
   },
 })