+++ packages/presets/src/theme/tokens/index.ts
@@ -2,6 +2,7 @@ import { defineTokens } from '@pandacss/dev'
 import { animations, easings } from './animations'
 import { colors } from './colors'
 import { lineHeights } from './line-heights'
+import { radii } from './radii'
 import { zIndex } from './z-index'
 
 export const tokens = defineTokens({
@@ -9,5 +10,6 @@ export const tokens = defineTokens({
   colors,
   easings,
   lineHeights,
+  radii,
   zIndex,
 })